fix: Separated user provider and delete token when logout
