+++ lib/src/features/user-profile/presentation/user_profile_screen.dart
@@ -2,12 +2,16 @@ import 'package:alcancia/src/resources/colors/colors.dart';
 import 'package:alcancia/src/shared/components/alcancia_action_dialog.dart';
 import 'package:alcancia/src/shared/components/alcancia_button.dart';
 import 'package:alcancia/src/shared/components/alcancia_toolbar.dart';
+import 'package:alcancia/src/shared/models/storage_item.dart';
 import 'package:alcancia/src/shared/provider/auth_service_provider.dart';
-import 'package:alcancia/src/shared/provider/user.dart';
+import 'package:alcancia/src/shared/provider/user_provider.dart';
+import 'package:alcancia/src/shared/services/storage_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:url_launcher/url_launcher.dart';
+
+import '../../../shared/models/user_model.dart';
 
 class UserProfileScreen extends ConsumerWidget {
   UserProfileScreen({Key? key}) : super(key: key);
@@ -76,44 +80,47 @@ class UserProfileScreen extends ConsumerWidget {
               ),
               Spacer(),
               Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: AlcanciaButton(
-                  foregroundColor: Colors.red,
-                  side: BorderSide(color: Colors.red),
-                  buttonText: "Borrar cuenta",
-                  fontSize: 18,
-                  padding: const EdgeInsets.only(
-                      left: 24.0, right: 24.0, top: 4.0, bottom: 4.0),
-                  onPressed: () async {
-                    await showDialog(
-                        context: context,
-                        builder: (BuildContext ctx) {
-                          return AlcanciaActionDialog(
-                              child: Text(
-                                "¿Seguro que quieres borrar tu cuenta?\nEsta acción no se puede deshacer.",
-                                style: Theme.of(context).textTheme.titleLarge,
-                              ),
-                              acceptText: "Confirmar",
-                              acceptColor: Colors.red,
-                              cancelText: "Cancelar",
-                              acceptAction: () async {
-                                try {
-                                  await authService.deleteAccount();
-                                  context.go("/");
-                                  ref.read(userProvider.notifier).setUser(null);
-                                } catch (e) {
-                                  ScaffoldMessenger.of(context).showSnackBar(_snackBar(context, "Hubo un problema al borrar tu cuenta."));
-                                }
-                              });
-                        });
-                  },
-                  rounded: true,
-                  icon: Padding(
-                    padding: EdgeInsets.all(8.0),
-                    child: Icon(Icons.delete_forever),
-                  ),
-                )
-              ),
+                  padding: const EdgeInsets.all(16.0),
+                  child: AlcanciaButton(
+                    foregroundColor: Colors.red,
+                    side: BorderSide(color: Colors.red),
+                    buttonText: "Borrar cuenta",
+                    fontSize: 18,
+                    padding: const EdgeInsets.only(
+                        left: 24.0, right: 24.0, top: 4.0, bottom: 4.0),
+                    onPressed: () async {
+                      await showDialog(
+                          context: context,
+                          builder: (BuildContext ctx) {
+                            return AlcanciaActionDialog(
+                                child: Text(
+                                  "¿Seguro que quieres borrar tu cuenta?\nEsta acción no se puede deshacer.",
+                                  style: Theme.of(context).textTheme.titleLarge,
+                                ),
+                                acceptText: "Confirmar",
+                                acceptColor: Colors.red,
+                                cancelText: "Cancelar",
+                                acceptAction: () async {
+                                  try {
+                                    await authService.deleteAccount();
+                                    context.go("/");
+                                    ref
+                                        .read(userProvider.notifier)
+                                        .setUser(null);
+                                  } catch (e) {
+                                    ScaffoldMessenger.of(context).showSnackBar(
+                                        _snackBar(context,
+                                            "Hubo un problema al borrar tu cuenta."));
+                                  }
+                                });
+                          });
+                    },
+                    rounded: true,
+                    icon: Padding(
+                      padding: EdgeInsets.all(8.0),
+                      child: Icon(Icons.delete_forever),
+                    ),
+                  )),
               Padding(
                   padding: const EdgeInsets.all(16.0),
                   child: AlcanciaButton(
@@ -138,10 +145,16 @@ class UserProfileScreen extends ConsumerWidget {
                                 acceptAction: () async {
                                   try {
                                     await authService.logout();
+                                    await deleteToken();
+                                    ref
+                                        .read(userProvider.notifier)
+                                        .setUser(null);
+
                                     context.go("/");
-                                    ref.read(userProvider.notifier).setUser(null);
                                   } catch (e) {
-                                    ScaffoldMessenger.of(context).showSnackBar(_snackBar(context, "Hubo un problema al cerrar sesión"));
+                                    ScaffoldMessenger.of(context).showSnackBar(
+                                        _snackBar(context,
+                                            "Hubo un problema al cerrar sesión"));
                                   }
                                 });
                           });
@@ -151,8 +164,7 @@ class UserProfileScreen extends ConsumerWidget {
                       padding: EdgeInsets.all(8.0),
                       child: Icon(Icons.delete_forever),
                     ),
-                  )
-              ),
+                  )),
             ],
           ),
         ),
@@ -225,5 +237,10 @@ class UserProfileScreen extends ConsumerWidget {
     if (!await launchUrl(url)) {
       throw 'Could not launch $url';
     }
+  }
+
+  deleteToken() async {
+    final StorageService _storageService = StorageService();
+    await _storageService.deleteSecureData("token");
   }
 }
