+++ lib/src/shared/components/dashboard/navbar.dart
@@ -1,6 +1,6 @@
 import 'package:alcancia/src/shared/components/alcancia_components.dart';
 import 'package:alcancia/src/shared/graphql/queries.dart';
-import 'package:alcancia/src/shared/provider/user.dart';
+import 'package:alcancia/src/shared/provider/user_provider.dart';
 import 'package:alcancia/src/shared/services/storage_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
