+++ lib/src/shared/provider/router_provider.dart
@@ -17,6 +17,7 @@ import 'package:graphql_flutter/graphql_flutter.dart';
 Future<bool> isUserAuthenticated() async {
   StorageService service = StorageService();
   var token = await service.readSecureData("token");
+  print("TOKEN: ${token}");
   GraphQLConfig graphQLConfiguration = GraphQLConfig(token: "${token}");
   GraphQLClient client = graphQLConfiguration.clientToQuery();
   var result = await client.query(QueryOptions(document: gql(isAuthenticated)));
@@ -103,6 +104,7 @@ final routerProvider = Provider<GoRouter>(
           }
           // return "/dashboard";
         }
+        return null;
         // means is not authenticated
         // return state.location != "/login" ? "/login" : null;
       },
